+++ printer/options.go
@@ -6,4 +6,7 @@ type Options struct {
 	// ShowIcons toggles display of the
 	// icons on the end Graph
 	ShowIcons bool
+
+	// AlternativeNodeNames will use Node names instead of canonical names
+	AlternativeNodeNames bool
 }